Crate frunk_laws

Frunk Laws

This library contains laws that can be used to test the implementation of algebras declared in Frunk

Modules

monoid_laws

Module that holds laws for Monoid implementations

semigroup_laws

Module that holds laws for Semigroup implementations

wrapper

This module holds the Wrapper newtype; used to write instances of typeclasses that we don’t define for types we don’t own